+++ fiddles/angular2/fiddle-0025-TreeView/src/meta.ts
@@ -1,0 +1,8 @@
+
+var metadata = {
+    fiddleHeader: 'Angular 2 - TreeView',
+    urls: {
+        github: 'https://github.com/bradyhouse/house/tree/master/fiddles/angular2/fiddle-0025-TreeView'
+    },
+    consoleTag: 'H O U S E ~ f i d d l e s'
+};